utils split across multiple files; internal and external imports working; utils tests split across multiple files; tests working locally
