+++ supervision/__init__.py
@@ -46,43 +46,51 @@ from supervision.detection.line_zone import (
     LineZoneAnnotator,
     LineZoneAnnotatorMulticlass,
 )
-from supervision.detection.overlap_filter import (
+from supervision.detection.utils.iou_and_nms import (
+    box_iou,
+    box_iou_batch,
+    box_iou_batch_with_jaccard,
+    oriented_box_iou_batch,
+    mask_iou_batch,
     OverlapFilter,
+    OverlapMetric,
     box_non_max_merge,
     box_non_max_suppression,
     mask_non_max_suppression,
+)
+from supervision.detection.utils.masks import (
+    move_masks,
+    calculate_masks_centroids,
+    contains_holes,
+    contains_multiple_segments
+)
+from supervision.detection.utils.converters import (
+    xcycwh_to_xyxy,
+    xywh_to_xyxy,
+    xyxy_to_polygons,
+    xyxy_to_xcycarh,
+    xyxy_to_xywh,
+    mask_to_polygons,
+    mask_to_xyxy,
+    polygon_to_mask,
+    polygon_to_xyxy,
+)
+from supervision.detection.utils.boxes import (
+    clip_boxes,
+    denormalize_boxes,
+    move_boxes,
+    pad_boxes,
+    scale_boxes,
+)
+from supervision.detection.utils.polygons import (
+    filter_polygons_by_area,
+    approximate_polygon
 )
 from supervision.detection.tools.csv_sink import CSVSink
 from supervision.detection.tools.inference_slicer import InferenceSlicer
 from supervision.detection.tools.json_sink import JSONSink
 from supervision.detection.tools.polygon_zone import PolygonZone, PolygonZoneAnnotator
 from supervision.detection.tools.smoother import DetectionsSmoother
-from supervision.detection.utils import (
-    box_iou,
-    box_iou_batch,
-    box_iou_batch_with_jaccard,
-    calculate_masks_centroids,
-    clip_boxes,
-    contains_holes,
-    contains_multiple_segments,
-    denormalize_boxes,
-    filter_polygons_by_area,
-    mask_iou_batch,
-    mask_to_polygons,
-    mask_to_xyxy,
-    move_boxes,
-    move_masks,
-    oriented_box_iou_batch,
-    pad_boxes,
-    polygon_to_mask,
-    polygon_to_xyxy,
-    scale_boxes,
-    xcycwh_to_xyxy,
-    xywh_to_xyxy,
-    xyxy_to_polygons,
-    xyxy_to_xcycarh,
-    xyxy_to_xywh,
-)
 from supervision.detection.vlm import LMM, VLM
 from supervision.draw.color import Color, ColorPalette
 from supervision.draw.utils import (
@@ -167,6 +175,7 @@ __all__ = [
     "MeanAveragePrecision",
     "OrientedBoxAnnotator",
     "OverlapFilter",
+    "OverlapMetric",
     "PercentageBarAnnotator",
     "PixelateAnnotator",
     "Point",
@@ -235,4 +244,5 @@ __all__ = [
     "xyxy_to_polygons",
     "xyxy_to_xcycarh",
     "xyxy_to_xywh",
+    "approximate_polygon"
 ]
