+++ supervision/dataset/formats/coco.py
@@ -13,11 +13,9 @@ from supervision.dataset.utils import (
     rle_to_mask,
 )
 from supervision.detection.core import Detections
-from supervision.detection.utils import (
-    contains_holes,
-    contains_multiple_segments,
-    polygon_to_mask,
-)
+from supervision.detection.utils.converters import polygon_to_mask
+from supervision.detection.utils.masks import contains_holes, contains_multiple_segments
+
 from supervision.utils.file import read_json_file, save_json_file
 
 if TYPE_CHECKING:
