+++ supervision/keypoint/annotators.py
@@ -7,7 +7,7 @@ import cv2
 import numpy as np
 
 from supervision.annotators.base import ImageType
-from supervision.detection.utils import pad_boxes, spread_out_boxes
+from supervision.detection.utils.boxes import pad_boxes, spread_out_boxes
 from supervision.draw.color import Color
 from supervision.draw.utils import draw_rounded_rectangle
 from supervision.geometry.core import Rect
