+++ supervision/metrics/recall.py
@@ -9,11 +9,9 @@ from matplotlib import pyplot as plt
 
 from supervision.config import ORIENTED_BOX_COORDINATES
 from supervision.detection.core import Detections
-from supervision.detection.utils import (
-    box_iou_batch,
-    mask_iou_batch,
-    oriented_box_iou_batch,
-)
+from supervision.detection.utils.iou_and_nms import box_iou_batch, mask_iou_batch, \
+    oriented_box_iou_batch
+
 from supervision.draw.color import LEGACY_COLOR_PALETTE
 from supervision.metrics.core import AveragingMethod, Metric, MetricTarget
 from supervision.metrics.utils.object_size import (
