+++ test/detection/utils/test_iou_and_nms.py
@@ -5,12 +5,8 @@ from contextlib import ExitStack as DoesNotRaise
 import numpy as np
 import pytest
 
-from supervision.detection.overlap_filter import (
-    box_non_max_suppression,
-    group_overlapping_boxes,
-    mask_non_max_merge,
-    mask_non_max_suppression,
-)
+from supervision.detection.utils.iou_and_nms import box_non_max_suppression, \
+    mask_non_max_suppression, mask_non_max_merge, _group_overlapping_boxes
 
 
 @pytest.mark.parametrize(
@@ -133,7 +129,7 @@ def test_group_overlapping_boxes(
     exception: Exception,
 ) -> None:
     with exception:
-        result = group_overlapping_boxes(
+        result = _group_overlapping_boxes(
             predictions=predictions, iou_threshold=iou_threshold
         )
